moved calibration code to resources.
